refactor: reorganize database infrastructure following Clean Architecture
- Move storage/database.rs to database/database.rs
- Move storage/daos/ to database/daos/
- Move storage/migrations/ to database/migrations/
- Move storage/seeders/ to database/seeders/
- Update all imports to use new database module structure
- Remove storage/integration_test.rs (functionality moved to unit tests)
- Clean up storage module to only contain file-based storage if needed

This completes the Clean Architecture separation by creating a dedicated
database infrastructure module separate from general storage concerns.
