+++ src/infrastructure/database/daos/challenge_dao.rs
