+++ src/infrastructure/mod.rs
@@ -1,5 +1,6 @@
 pub mod cache;
 pub mod config;
+pub mod database;
 pub mod external;
 pub mod git;
 pub mod http;
