--- src/infrastructure/storage/mod.rs
@@ -1,10 +1,0 @@
-pub mod daos;
-pub mod database;
-pub mod migrations;
-pub mod seeders;
-
-#[cfg(test)]
-pub mod integration_test;
-
-pub use daos::*;
-pub use database::{Database, HasDatabase};
